+++ quantestpy/__init__.py
@@ -3,4 +3,4 @@ from .simulator.state_vector_circuit import StateVectorCircuit
 from .simulator.pauli_circuit import PauliCircuit
 from .assertion.get_ctrl_val import assert_get_ctrl_val
 from .assertion.get_tgt_val import assert_get_tgt_val
-from .assertion.not_entangled import assert_not_entangled
+from .assertion.assert_not_entangled import assert_not_entangled
